fix after 2nd review
